backend/requirements.txt
django>=3.2.8
djangorestframework>=3.12.4
